feat: add openebs_upgrade (#3826)
* feat: add openebs_upgrade

Signed-off-by: Niladri Halder <niladri.halder26@gmail.com>

* feat(upgrade-job): install CRDs intelligently
When upgrading from OpenEBS v3.x to v4.y, if the labels for the yaml
keys for release name, etc. have a different value, the installation
fails. This commit checks for CRDs and if the ones from the openebs/
openebs chart or its openebs flavoured sub-charts exist, then their
installation is skipped by means of helm chart values toggles.

Signed-off-by: Niladri Halder <niladri.halder26@gmail.com>

* chore(deps): update mayastor-extensions git submodule

Signed-off-by: Niladri Halder <niladri.halder26@gmail.com>

---------

Signed-off-by: Niladri Halder <niladri.halder26@gmail.com>